requirements.txt
Flask==3.0.3
gunicorn==22.0.0
requests==2.32.3
pymongo==4.8.0
python-dotenv==1.0.1
mongomock==4.1.2
